requirements.txt
influxdb_client==1.36.0
paho_mqtt==1.6.1
pandas==1.5.2
requests==2.28.1
aiocoap==0.4.5
influxdb_client==1.35.0
numpy==1.23.5
pandas==1.5.2
twilio==7.16.2
scikit-learn==1.2.1